WebCore: dataTransfer.types() should not return Files if file list is empty in the clipboard. https://bugs.webkit.org/show_bug.cgi?id=28891
Reviewed by David Levin.

We change the behavior to handle the empty file list in order to match the spec.

Tested by clipboard-file-access.html.

* platform/mac/ClipboardMac.mm:
(WebCore::addHTMLClipboardTypesForCocoaType):
(WebCore::ClipboardMac::types):

LayoutTests: dataTransfer.types() should not return Files if file list is empty in the clipboard.
https://bugs.webkit.org/show_bug.cgi?id=28891

Reviewed by David Levin.

Update the test script and expected result to reflect the behavior change.

* http/tests/security/clipboard/clipboard-file-access-expected.txt:
* http/tests/security/clipboard/resources/clipboard-file-access.js:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@48169 268f45cc-cd09-0410-ab3c-d52691b4dbfc

Conflicts:

	LayoutTests/ChangeLog
	LayoutTests/http/tests/security/clipboard/resources/clipboard-file-access.js
	WebCore/ChangeLog