MSAA: Accessibility of headings is not correct
https://bugs.webkit.org/show_bug.cgi?id=30937

Reviewed by Alice Liu.

WebCore:

* accessibility/AccessibilityObject.h:
(WebCore::AccessibilityObject::stringRoleForMSAA):
(WebCore::AccessibilityObject::descriptionForMSAA):

* accessibility/AccessibilityRenderObject.cpp:
(WebCore::shouldReturnTagNameAsRoleForMSAA):
If the element's tag name is one of h1, h2, h3, h4, h5, h6, return
true.
(WebCore::AccessibilityRenderObject::stringRoleForMSAA):
If the element should return its tag name as the role, return the tag
name.
(WebCore::AccessibilityRenderObject::positionalDescriptionForMSAA):
If the object is a heading, return the string "L" followed by the
heading level.
(WebCore::AccessibilityRenderObject::descriptionForMSAA):
If the object has a positional description, return it. Otherwise, get
the accessibility description, and prefix it with "Description" so that
MSAA clients know that it's not a positional description.

* accessibility/AccessibilityRenderObject.h:

WebKit/win:

* AccessibleBase.cpp:
(AccessibleBase::get_accDescription):
Call the object's descriptionForMSAA(). Moved the comment to the
WebCore file.
(AccessibleBase::get_accRole):
If the object has a string role, return that. Otherwise, return the
integer role.

* AccessibleBase.h:
Removed description(), as this was moved to WebCore.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@50354 268f45cc-cd09-0410-ab3c-d52691b4dbfc