[Qt] Add automatic test case for plugins that use QWidget
https://bugs.webkit.org/show_bug.cgi?id=31203

Change the RenderTreeAsText to look into the WebCore::Widget
and print the state of the platformWidget. Change the DRT to
use QtUiTools to be able to initiate QProgressBar and such
as a plugin. Add the test case, add the result, update the
other results due the new test.

WebCore:
* rendering/RenderTreeAsText.cpp:
(WebCore::operator<<): Add special case for RenderPart

WebKitTools:
* DumpRenderTree/qt/DumpRenderTree.cpp:
(WebCore::WebPage::createPlugin):
* DumpRenderTree/qt/DumpRenderTree.h:
* DumpRenderTree/qt/DumpRenderTree.pro:

LayoutTest:
* platform/gtk/Skipped:
* platform/mac/Skipped:
* platform/qt/plugins/netscape-dom-access-expected.txt: Added
* platform/qt/plugins/qt-qwidget-plugin-expected.txt: Added
* platform/win/Skipped:
* plugins/qt-qwidget-plugin.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@51410 268f45cc-cd09-0410-ab3c-d52691b4dbfc