repaint events from outside the viewport aren't received https://bugs.webkit.org/show_bug.cgi?id=32081
Reviewed by Simon Fraser.

When using a tiled backing store for painting, you need to receive
event from outside the viewport. Setting the viewport to the size
of the contents is not an option if you want to make use of WebCore's
infrastructure for drawing scrollbars etc.

A new property, paintsEntireContents, has been introduced for the
above use-case. It is settable, as tiling will be optional for Qt,
and for the not yet upstreamed EFL port, there will be two different
views, where only one of them are tiled.

No change in behavior, so no new tests added.

* page/FrameView.cpp:
(WebCore::FrameView::repaintContentRectangle):
* platform/ScrollView.cpp:
(WebCore::ScrollView::ScrollView):
(WebCore::ScrollView::setPaintsEntireContents):
(WebCore::ScrollView::wheelEvent):
* platform/ScrollView.h:
(WebCore::ScrollView::paintsEntireContents):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@51636 268f45cc-cd09-0410-ab3c-d52691b4dbfc