[DnD] effectAllowed and dropEffect can be set to bogus values. Fixes <https://bugs.webkit.org/show_bug.cgi?id=33635>.
Reviewed by Oliver Hunt.

WebCore: 

Test to make aure dropEffect and effectAllowed are being set to valid values
when they are being set (list of valid values given by HTML5 specification).
        
Also, drive by change to initialize dropEffect to none (as described in spec).

Test: fast/events/bogus-dropEffect-effectAllowed.html

* dom/Clipboard.cpp:
(WebCore::Clipboard::Clipboard): Initialize m_dropEffect to "none".
(WebCore::Clipboard::setDropEffect): Check if dropEffect is being set to a valid value.
(WebCore::Clipboard::setEffectAllowed): Check if effectAllowed is being set to a valid value.

LayoutTests: 

Added a test to check the handling of setting effectAllowed and
dropEffect to bogus values (that it gets ignored), and updated results
to drag-and-drop because it uses a dummy value.

* fast/events/bogus-dropEffect-effectAllowed-expected.txt: Added.
* fast/events/bogus-dropEffect-effectAllowed.html: Added.
* fast/events/drag-and-drop-expected.txt:
* fast/events/drag-and-drop.html:



git-svn-id: http://svn.webkit.org/repository/webkit/trunk@53270 268f45cc-cd09-0410-ab3c-d52691b4dbfc